@@ -5,30 +5,26 @@ import com.liaoyb.qingqing.uaa.domain.User;
 import com.liaoyb.qingqing.uaa.repository.UserRepository;
 
 import org.apache.commons.lang3.RandomStringUtils;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Locale;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 
 /**
- * Test class for DomainUserDetailsService.
- *
- * @see DomainUserDetailsService
+ * Integrations tests for {@link DomainUserDetailsService}.
  */
-@RunWith(SpringRunner.class)
 @SpringBootTest(classes = UaaApp.class)
 @Transactional
-public class DomainUserDetailsServiceIntTest {
+public class DomainUserDetailsServiceIT {
 
     private static final String USER_ONE_LOGIN = "test-user-one";
     private static final String USER_ONE_EMAIL = "test-user-one@localhost";
@@ -47,7 +43,7 @@ public class DomainUserDetailsServiceIntTest {
     private User userTwo;
     private User userThree;
 
-    @Before
+    @BeforeEach
     public void init() {
         userOne = new User();
         userOne.setLogin(USER_ONE_LOGIN);
@@ -104,10 +100,11 @@ public class DomainUserDetailsServiceIntTest {
         assertThat(userDetails.getUsername()).isEqualTo(USER_TWO_LOGIN);
     }
 
-    @Test(expected = UsernameNotFoundException.class)
+    @Test
     @Transactional
     public void assertThatUserCanNotBeFoundByEmailIgnoreCase() {
-    domainUserDetailsService.loadUserByUsername(USER_TWO_EMAIL.toUpperCase(Locale.ENGLISH));
+        assertThatExceptionOfType(UsernameNotFoundException.class).isThrownBy(
+            () -> domainUserDetailsService.loadUserByUsername(USER_TWO_EMAIL.toUpperCase(Locale.ENGLISH)));
     }
 
     @Test
@@ -118,10 +115,11 @@ public class DomainUserDetailsServiceIntTest {
         assertThat(userDetails.getUsername()).isEqualTo(USER_ONE_LOGIN);
     }
 
-    @Test(expected = UserNotActivatedException.class)
+    @Test
     @Transactional
     public void assertThatUserNotActivatedExceptionIsThrownForNotActivatedUsers() {
-        domainUserDetailsService.loadUserByUsername(USER_THREE_LOGIN);
+        assertThatExceptionOfType(UserNotActivatedException.class).isThrownBy(
+            () -> domainUserDetailsService.loadUserByUsername(USER_THREE_LOGIN));
     }
 
 }
