@@ -1,12 +1,12 @@
 package com.liaoyb.qingqing.uaa.service;
+
 import com.liaoyb.qingqing.uaa.config.Constants;
 
 import com.liaoyb.qingqing.uaa.UaaApp;
 import com.liaoyb.qingqing.uaa.domain.User;
 import io.github.jhipster.config.JHipsterProperties;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
 import org.mockito.MockitoAnnotations;
@@ -16,7 +16,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.MessageSource;
 import org.springframework.mail.MailSendException;
 import org.springframework.mail.javamail.JavaMailSenderImpl;
-import org.springframework.test.context.junit4.SpringRunner;
 import org.thymeleaf.spring5.SpringTemplateEngine;
 
 import javax.mail.Multipart;
@@ -29,9 +28,11 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
-@RunWith(SpringRunner.class)
+/**
+ * Integration tests for {@link MailService}.
+ */
 @SpringBootTest(classes = UaaApp.class)
-public class MailServiceIntTest {
+public class MailServiceIT {
 
     @Autowired
     private JHipsterProperties jHipsterProperties;
@@ -50,7 +51,7 @@ public class MailServiceIntTest {
 
     private MailService mailService;
 
-    @Before
+    @BeforeEach
     public void setup() {
         MockitoAnnotations.initMocks(this);
         doNothing().when(javaMailSender).send(any(MimeMessage.class));
