@@ -9,8 +9,8 @@ import java.util.*;
 
 /**
  * Utility class to load a Spring profile to be used as default
- * when there is no <code>spring.profiles.active</code> set in the environment or as command line argument.
- * If the value is not available in <code>application.yml</code> then <code>dev</code> profile will be used as default.
+ * when there is no {@code spring.profiles.active} set in the environment or as command line argument.
+ * If the value is not available in {@code application.yml} then {@code dev} profile will be used as default.
  */
 public final class DefaultProfileUtil {
 
@@ -22,30 +22,16 @@ public final class DefaultProfileUtil {
     /**
      * Set a default to use when no profile is configured.
      *
-     * @param app the Spring application
+     * @param app the Spring application.
      */
     public static void addDefaultProfile(SpringApplication app) {
         Map<String, Object> defProperties = new HashMap<>();
         /*
         * The default profile to use when no other profiles are defined
-        * This cannot be set in the <code>application.yml</code> file.
+        * This cannot be set in the application.yml file.
         * See https://github.com/spring-projects/spring-boot/issues/1219
         */
         defProperties.put(SPRING_PROFILE_DEFAULT, JHipsterConstants.SPRING_PROFILE_DEVELOPMENT);
         app.setDefaultProperties(defProperties);
-    }
-
-    /**
-     * Get the profiles that are applied else get default profiles.
-     *
-     * @param env spring environment
-     * @return profiles
-     */
-    public static String[] getActiveProfiles(Environment env) {
-        String[] profiles = env.getActiveProfiles();
-        if (profiles.length == 0) {
-            return env.getDefaultProfiles();
-        }
-        return profiles;
     }
 }
