@@ -5,7 +5,7 @@ import com.liaoyb.qingqing.uaa.domain.Authority;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 /**
- * Spring Data JPA repository for the Authority entity.
+ * Spring Data JPA repository for the {@link Authority} entity.
  */
 public interface AuthorityRepository extends JpaRepository<Authority, String> {
 }
