@@ -16,7 +16,7 @@ import java.time.Instant;
 import java.util.*;
 
 /**
- * An implementation of Spring Boot's AuditEventRepository.
+ * An implementation of Spring Boot's {@link AuditEventRepository}.
  */
 @Repository
 public class CustomAuditEventRepository implements AuditEventRepository {
