@@ -17,7 +17,7 @@ public final class SecurityUtils {
     /**
      * Get the login of the current user.
      *
-     * @return the login of the current user
+     * @return the login of the current user.
      */
     public static Optional<String> getCurrentUserLogin() {
         SecurityContext securityContext = SecurityContextHolder.getContext();
@@ -36,7 +36,7 @@ public final class SecurityUtils {
     /**
      * Check if a user is authenticated.
      *
-     * @return true if the user is authenticated, false otherwise
+     * @return true if the user is authenticated, false otherwise.
      */
     public static boolean isAuthenticated() {
         SecurityContext securityContext = SecurityContextHolder.getContext();
@@ -49,10 +49,10 @@ public final class SecurityUtils {
     /**
      * If the current user has a specific authority (security role).
      * <p>
-     * The name of this method comes from the isUserInRole() method in the Servlet API
+     * The name of this method comes from the {@code isUserInRole()} method in the Servlet API.
      *
-     * @param authority the authority to check
-     * @return true if the current user has the authority, false otherwise
+     * @param authority the authority to check.
+     * @return true if the current user has the authority, false otherwise.
      */
     public static boolean isCurrentUserInRole(String authority) {
         SecurityContext securityContext = SecurityContextHolder.getContext();
