@@ -14,7 +14,7 @@ import java.util.Optional;
 /**
  * Service for managing audit events.
  * <p>
- * This is the default implementation to support SpringBoot Actuator AuditEventRepository
+ * This is the default implementation to support SpringBoot Actuator {@code AuditEventRepository}.
  */
 @Service
 @Transactional
