@@ -21,7 +21,7 @@ import org.thymeleaf.spring5.SpringTemplateEngine;
 /**
  * Service for sending emails.
  * <p>
- * We use the @Async annotation to send emails asynchronously.
+ * We use the {@link Async} annotation to send emails asynchronously.
  */
 @Service
 public class MailService {
