@@ -9,8 +9,8 @@ import org.springframework.security.oauth2.provider.token.store.JwtAccessTokenCo
 import org.springframework.web.client.RestTemplate;
 
 /**
- * Overrides UAA specific beans, so they do not interfere the testing
- * This configuration must be included in @SpringBootTest in order to take effect.
+ * Overrides UAA specific beans, so they do not interfere the testing.
+ * This configuration must be included in {@code @SpringBootTest} in order to take effect.
  */
 @Configuration
 public class SecurityBeanOverrideConfiguration {
