@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 /**
- * Spring Data JPA repository for the DateTimeWrapper entity.
+ * Spring Data JPA repository for the {@link DateTimeWrapper} entity.
  */
 @Repository
 public interface DateTimeWrapperRepository extends JpaRepository<DateTimeWrapper, Long> {
