@@ -11,7 +11,6 @@ import org.springframework.format.support.FormattingConversionService;
 import org.springframework.http.MediaType;
 
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeParseException;
 
@@ -25,10 +24,7 @@ public final class TestUtil {
     private static final ObjectMapper mapper = createObjectMapper();
 
     /** MediaType for JSON UTF8 */
-    public static final MediaType APPLICATION_JSON_UTF8 = new MediaType(
-            MediaType.APPLICATION_JSON.getType(),
-            MediaType.APPLICATION_JSON.getSubtype(), StandardCharsets.UTF_8);
-
+    public static final MediaType APPLICATION_JSON_UTF8 = MediaType.APPLICATION_JSON_UTF8;
 
     private static ObjectMapper createObjectMapper() {
         ObjectMapper mapper = new ObjectMapper();
@@ -40,22 +36,20 @@ public final class TestUtil {
     /**
      * Convert an object to JSON byte array.
      *
-     * @param object
-     *            the object to convert
-     * @return the JSON byte array
+     * @param object the object to convert.
+     * @return the JSON byte array.
      * @throws IOException
      */
-    public static byte[] convertObjectToJsonBytes(Object object)
-            throws IOException {
+    public static byte[] convertObjectToJsonBytes(Object object) throws IOException {
         return mapper.writeValueAsBytes(object);
     }
 
     /**
      * Create a byte array with a specific size filled with specified data.
      *
-     * @param size the size of the byte array
-     * @param data the data to put in the byte array
-     * @return the JSON byte array
+     * @param size the size of the byte array.
+     * @param data the data to put in the byte array.
+     * @return the JSON byte array.
      */
     public static byte[] createByteArray(int size, String data) {
         byte[] byteArray = new byte[size];
@@ -99,8 +93,8 @@ public final class TestUtil {
     }
 
     /**
-     * Creates a matcher that matches when the examined string reprensents the same instant as the reference datetime
-     * @param date the reference datetime against which the examined string is checked
+     * Creates a matcher that matches when the examined string represents the same instant as the reference datetime.
+     * @param date the reference datetime against which the examined string is checked.
      */
     public static ZonedDateTimeMatcher sameInstant(ZonedDateTime date) {
         return new ZonedDateTimeMatcher(date);
@@ -126,8 +120,8 @@ public final class TestUtil {
     }
 
     /**
-     * Create a FormattingConversionService which use ISO date format, instead of the localized one.
-     * @return the FormattingConversionService
+     * Create a {@link FormattingConversionService} which use ISO date format, instead of the localized one.
+     * @return the {@link FormattingConversionService}.
      */
     public static FormattingConversionService createFormattingConversionService() {
         DefaultFormattingConversionService dfcs = new DefaultFormattingConversionService ();
